requirements.txt
requests
azure-functions
azure.identity
azure.keyvault.secrets